server/requirements.txt
flask
flask_cors
flask_compress
Pillow
numpy
flask-mongoengine
